// src/enums/eApgRegrTypes.ts
/** -----------------------------------------------------------------------
 * @module [REGR]
 * @author [APG] ANGELI Paolo Giusto
 * @version 0.8.0 [APG 2022/05/29] Porting to Deno 
 * -----------------------------------------------------------------------
 */
export enum eApgRegrTypes {
  LINEAR = "linear",
  POLYNOMIAL = "polynomial",
  POWER = "power",
  LOGARITMIC = "logaritmic",
  EXPONENTIAL = "exponential"
}