// src/interfaces/IApgRegrOptions.ts
/** -----------------------------------------------------------------------
 * @module [REGR]
 * @author [APG] ANGELI Paolo Giusto
 * @version 0.8.0 [APG 2022/05/29] Porting to Deno 
 * -----------------------------------------------------------------------
 */
export interface IApgRegrOptions {
  order: number;
  precision: number;
}